+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @douglasneuroinformatics/ui
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @douglasneuroinformatics/form-types@3.0.0
+
 ## 3.0.0
 
 ### Major Changes